requirements.txt
Django>=5.2.8
requests>=2.32.0
Pillow>=11.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
pytesseract>=0.3.13
pdf2image>=1.17.0
python-dateutil>=2.9.0
opencv-python>=4.8.0
gunicorn>=21.2.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
pytz>=2025.2

# Testing
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-factoryboy>=2.5.1
factory-boy>=3.3.0

# API REST
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-filter>=23.3
drf-spectacular>=0.27.0  # Para documentación Swagger/OpenAPI

# Tareas asíncronas
celery>=5.3.4
redis>=5.0.1  # Para Celery broker y cache

# Cache avanzado (opcional, con fallback)
django-redis>=5.4.0

